python_backend/requirements.txt
flask
motor
requests
asyncio
python-dotenv
flask[async]
pika
aio_pika